+++ examples/shared_samples/storage_system.rb
@@ -1,4 +1,4 @@
-# (C) Copyright 2016 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2017 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # You may not use this file except in compliance with the License.
@@ -9,27 +9,43 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-require_relative '../../_client' # Gives access to @client
+require_relative '../_client' # Gives access to @client
 
-# Example: Create a storage system
-# NOTE: This will create a storage system, then delete it.
 # NOTE: You'll need to add the following instance variables to the _client.rb file with valid credentials for your environment:
 #   @storage_system_ip
 #   @storage_system_username
 #   @storage_system_password
 
+# Supported APIs:
+# - API200 for C7000
+# - API300 for C7000
+# - API300 for Synergy
+# - API500 for C7000 (see /examples/api500/storage_system.rb)
+# - API500 for Synergy (see /examples/api500/storage_system.rb)
+
+# Resources that can be created according to parameters:
+# api_version = 200 & variant = any to OneviewSDK::API200::StorageSystem
+# api_version = 300 & variant = C7000 to OneviewSDK::API300::C7000::StorageSystem
+# api_version = 300 & variant = Synergy to OneviewSDK::API300::Synergy::StorageSystem
+
 raise 'ERROR: Must set @storage_system_ip in _client.rb' unless @storage_system_ip
 raise 'ERROR: Must set @storage_system_username in _client.rb' unless @storage_system_username
 raise 'ERROR: Must set @storage_system_password in _client.rb' unless @storage_system_password
 
-# Example: Create a storage system
+if @client.api_version == 500
+  raise "If you want execute sample for API #{@client.api_version}, you should execute the ruby file '/examples/api500/storage_system.rb'"
+end
+
+# Resource Class used in this sample
+storage_system_class = OneviewSDK.resource_named('StorageSystem', @client.api_version)
+
 options = {
   ip_hostname: @storage_system_ip,
   username: @storage_system_username,
   password: @storage_system_password
 }
 
-storage1 = OneviewSDK::API300::C7000::StorageSystem.new(@client)
+storage1 = storage_system_class.new(@client)
 storage1['credentials'] = options
 storage1['managedDomain'] = 'TestDomain'
 puts "\nAdding a storage system with Managed Domain: #{storage1['managedDomain']}"
@@ -38,11 +54,11 @@ storage1.add
 puts "\nStorage system added successfully."
 
 puts "\nFinding a storage system with hostname: #{options[:ip_hostname]}"
-OneviewSDK::API300::C7000::StorageSystem.find_by(@client, credentials: { ip_hostname: options[:ip_hostname] }).each do |storage|
+storage_system_class.find_by(@client, credentials: { ip_hostname: options[:ip_hostname] }).each do |storage|
   puts "\nStorage system with name #{storage['name']} found."
 end
 
-storage2 = OneviewSDK::API300::C7000::StorageSystem.new(@client, credentials: { ip_hostname: options[:ip_hostname] })
+storage2 = storage_system_class.new(@client, credentials: { ip_hostname: options[:ip_hostname] })
 storage2.retrieve!
 puts "\nRefreshing a storage system with #{options[:ip_hostname]}"
 puts "and state #{storage2['refreshState']}"